requirements.txt
flask
pyjwt
bcrypt
requests